sdk: refactor: split channels epics into submodules
It was getting too big.
