+++ raiden-ts/package.json
@@ -7,21 +7,23 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "prepare": "yarn run versions && yarn run contracts",
-    "build": "yarn prepare && tsc && tsc -p ./tsconfig.cjs.json && yarn run contracts:post",
+    "build": "yarn clean:build && yarn prepare && tsc && tsc -p ./tsconfig.cjs.json && yarn run contracts:post",
     "test": "yarn lint && NODE_ENV=development jest --coverage --testPathIgnorePatterns tests/e2e",
     "test:unit": "NODE_ENV=development jest --coverage tests/unit",
     "test:integration": "NODE_ENV=development jest --coverage tests/integration",
     "test:e2e": "NODE_ENV=development jest --coverage tests/e2e",
     "test:e2e:docker": "bash ../e2e-environment/run-e2e-tests.sh",
     "lint": "eslint --max-warnings 0 --ext '.ts' .",
-    "clean": "rimraf dist/ dist:cjs/ .coverage/ src/contracts/ src/abi/ src/deployment/ src/versions.json docs/",
+    "clean:build": "rimraf dist/ dist:cjs/",
+    "clean:rest": "rimraf .coverage/ src/contracts/ src/abi/ src/deployment/ src/versions.json docs/",
+    "clean": "yarn clean:build && yarn clean:rest",
     "contracts:pre": "node ./scripts/copyContracts.js prebuild",
     "contracts:build": "typechain --target=ethers-v5 'src/abi/*.json' --out-dir src/contracts",
     "contracts:post": "node ./scripts/copyContracts.js postbuild",
     "contracts": "yarn run contracts:pre && yarn run contracts:build",
     "versions": "node ./scripts/versions.js",
     "docs": "typedoc",
-    "prepack": "yarn clean && yarn build && yarn run docs",
+    "prepack": "yarn clean:rest && yarn build && yarn run docs",
     "update-submodules": "git submodule sync && git submodule update --init"
   },
   "repository": {
